+++ CHANGELOG.md
@@ -1,11 +1,15 @@
 # Change log for Microsoft365DSC
+
 # UNRELEASED
 
-* AADGroupEligibilitySchedule
-  * FIXES [#5741] Missing -All variable which prevented enumeration of more than 100 Groups
 * AADRoleEligibilityScheduleRequest
   * Reduce call count when reconciling object type
     FIXES [#5621](https://github.com/microsoft/Microsoft365DSC/issues/5621)
+
+# 1.25.212.2
+
+* MISC
+  * Fixing connection errors in Power Platforms resources.
 
 # 1.25.212.1
 
@@ -17,6 +21,8 @@
 * AADCrossTenantAccessPolicyConfigurationPartner
   * Fixes an issue with CIM class export.
     FIXES [#5711](https://github.com/microsoft/Microsoft365DSC/issues/5711)
+* AADGroupEligibilitySchedule
+  * FIXES [#5741] Missing -All variable which prevented enumeration of more than 100 Groups
 * EXORetentionPolicyTag
   * Initial release.
 * EXOSmtpDaneInbound
@@ -27,6 +33,7 @@
   * Simplify export behavior for all resources and complex objects.
 * DEPENDENCIES
   * Removed dependency on Microsoft.PowerApps.Administration.PowerShell.
+  * Updated DSCParser to version 2.0.0.15.
   * Updated MSCloudLoginAssistant to version 1.1.37.
   * Updated ReverseDSC to version 2.0.0.25.